% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wordcloud.R
\name{generate_wordcloud}
\alias{generate_wordcloud}
\title{generate_wordcloud}
\usage{
generate_wordcloud(
  .file_name = "Interactive wordcloud (Responses)",
  .file_path = "temp_wordcloud_responses.csv",
  .column_name = "What is your favourite colour?",
  .stopwords = "english",
  .removewords = NULL,
  .return = "wordcloud",
  ...
)
}
\arguments{
\item{.file_name}{A chr string of the name of the survey response google sheet file (as in the Google Drive)}

\item{.file_path}{A chr string of the file location the google sheet will be downloaded to locally. (Needs ".csv")}

\item{.column_name}{A chr string denoting the question/column name of the responses to be integrated into the wordcloud.}

\item{.stopwords}{Options for cleaning common words. For further options, see ?tm::stopwords}

\item{.removewords}{Null, or a character string vector containing additional words to remove}

\item{.return}{A chr string, either "wordcloud" for plotting wordcloud output, or "data" for outputting the frequency table.}

\item{...}{Further arguments passed to wordcloud2::wordcould2()}
}
\value{
Either a wordcloud plot, or a frequency table
}
\description{
A function to download data, clean and prepare, and plot the wordcloud, from the specified file and question.
}
\examples{
## not run
# generate_wordcloud(size = 0.5)
# generate_wordcloud(.return = "data")
}
